+++ lib/src/core/utils/formats/date_formats.dart
@@ -2,31 +2,10 @@ import 'package:cloud_firestore/cloud_firestore.dart' show Timestamp;
 import 'package:intl/intl.dart' as df show DateFormat;
 
 class DateFormat {
-  /// Retorna tempo de duração fazendo a comparação do
-  /// tempo final com o tempo passado
-  /// EXEMPLOS DE RETORNO: 1 Dia, 30 min, 2 Hrs, 1 H, 3 Dias, ...
-  /// Necessário passar o DateTime.
-  static String permanenceTime({required DateTime lastTime}) {
-    final time = DateTime.now().difference(lastTime);
-    if (time.inSeconds < 60) {
-      return '${time.inSeconds} s';
-    } else if (time.inMinutes < 60) {
-      return '${time.inMinutes} min';
-    } else if (time.inHours < 24) {
-      return '${time.inHours} h';
-    } else if (time.inDays > 0 && time.inDays == 1) {
-      return '${time.inDays} Dia';
-    } else if (time.inDays > 1) {
-      return '${time.inDays} Dias';
-    } else {
-      return 'Não informado';
-    }
-  }
-
   static String toDate(
     DateTime? date, {
     String pattern = 'dd/MM/yyyy',
-    String locale = 'pt_BR',
+    String? locale = 'pt_BR',
   }) {
     if (date == null) return '';
 
@@ -36,55 +15,77 @@ class DateFormat {
   static DateTime tryParseOrDateNow(
     String? date, {
     String pattern = 'yyyy-MM-dd',
+    String? locale = 'pt_BR',
   }) {
     if (date == null || date.isEmpty) {
       return DateTime.now();
     } else {
-      return tryParse(date, pattern: pattern) ?? DateTime.now();
+      return tryParse(date, pattern: pattern, locale: locale) ?? DateTime.now();
     }
   }
 
-  static DateTime? tryParse(String? date, {String pattern = 'yyyy-MM-dd'}) {
+  static DateTime? tryParse(
+    String? date, {
+    String pattern = 'yyyy-MM-dd',
+    String? locale = 'pt_BR',
+  }) {
     if (date == null || date.isEmpty) {
       return null;
     } else {
       try {
-        return df.DateFormat(pattern).parse(date).toLocal();
+        return df.DateFormat(pattern, locale).parse(date).toLocal();
       } catch (e) {
         return null;
       }
     }
   }
 
-  static String toTime(DateTime? date, {String pattern = 'HH:mm:ss'}) {
+  static String toTime(
+    DateTime? date, {
+    String pattern = 'HH:mm:ss',
+    String? locale = 'pt_BR',
+  }) {
     if (date == null) return '';
 
-    return df.DateFormat(pattern, 'pt_BR').format(date.toLocal());
+    return df.DateFormat(pattern, locale).format(date.toLocal());
   }
 
   static String toDateTime(
     DateTime? date, {
     String pattern = 'dd/MM/yyyy HH:mm:ss',
+    String? locale = 'pt_BR',
   }) {
     if (date == null) return '';
 
-    return df.DateFormat(pattern, 'pt_BR').format(date.toLocal());
+    return df.DateFormat(pattern, locale).format(date.toLocal());
   }
 
-  static DateTime? fromString(String date, {format = ''}) {
+  static DateTime? fromString(
+    String date, {
+    format = '',
+    String? locale = 'pt_BR',
+  }) {
     try {
-      return df.DateFormat(format).parse(date).toLocal();
+      return df.DateFormat(format, locale).parse(date).toLocal();
     } catch (e) {
       return null;
     }
   }
 
-  static bool isSameDate(DateTime date1, DateTime date2) {
-    return toDate(date1) == toDate(date2);
+  static bool isSameDate(
+    DateTime date1,
+    DateTime date2, {
+    String? locale = 'pt_BR',
+  }) {
+    return toDate(date1, locale: locale) == toDate(date2, locale: locale);
   }
 
-  static bool dateIsToday(DateTime date) {
-    return toDate(date) == toDate(DateTime.now());
+  static bool dateIsToday(DateTime date, {String? locale = 'pt_BR'}) {
+    return toDate(date, locale: locale) ==
+        toDate(
+          DateTime.now(),
+          locale: locale,
+        );
   }
 
   static DateTime? timestampToDate(Timestamp? date) {