+++ lib/src/core/utils/period_type.dart
@@ -1,94 +1,94 @@
-class PeriodTypeUtils {
-  static String returnMonthByNumber(String number) {
-    switch (number) {
-      case '0':
-        return 'Selecione o período';
-      case '1':
-        return 'Janeiro';
-      case '2':
-        return 'Fevereiro';
-      case '3':
-        return 'Março';
-      case '4':
-        return 'Abril';
-      case '5':
-        return 'Maio';
-      case '6':
-        return 'Junho';
-      case '7':
-        return 'Julho';
-      case '8':
-        return 'Agosto';
-      case '9':
-        return 'Setembro';
-      case '10':
-        return 'Outubro';
-      case '11':
-        return 'Novembro';
-      case '12':
-        return 'Dezembro';
-      default:
-        return '';
-    }
-  }
+// class PeriodTypeUtils {
+//   static String returnMonthByNumber(String number) {
+//     switch (number) {
+//       case '0':
+//         return 'Selecione o período';
+//       case '1':
+//         return 'Janeiro';
+//       case '2':
+//         return 'Fevereiro';
+//       case '3':
+//         return 'Março';
+//       case '4':
+//         return 'Abril';
+//       case '5':
+//         return 'Maio';
+//       case '6':
+//         return 'Junho';
+//       case '7':
+//         return 'Julho';
+//       case '8':
+//         return 'Agosto';
+//       case '9':
+//         return 'Setembro';
+//       case '10':
+//         return 'Outubro';
+//       case '11':
+//         return 'Novembro';
+//       case '12':
+//         return 'Dezembro';
+//       default:
+//         return '';
+//     }
+//   }
 
-  static String getMonthByInt(int number) {
-    switch (number) {
-      case 1:
-        return 'Janeiro';
-      case 2:
-        return 'Fevereiro';
-      case 3:
-        return 'Março';
-      case 4:
-        return 'Abril';
-      case 5:
-        return 'Maio';
-      case 6:
-        return 'Junho';
-      case 7:
-        return 'Julho';
-      case 8:
-        return 'Agosto';
-      case 9:
-        return 'Setembro';
-      case 10:
-        return 'Outubro';
-      case 11:
-        return 'Novembro';
-      case 12:
-        return 'Dezembro';
-      default:
-        return '';
-    }
-  }
+//   static String getMonthByInt(int number) {
+//     switch (number) {
+//       case 1:
+//         return 'Janeiro';
+//       case 2:
+//         return 'Fevereiro';
+//       case 3:
+//         return 'Março';
+//       case 4:
+//         return 'Abril';
+//       case 5:
+//         return 'Maio';
+//       case 6:
+//         return 'Junho';
+//       case 7:
+//         return 'Julho';
+//       case 8:
+//         return 'Agosto';
+//       case 9:
+//         return 'Setembro';
+//       case 10:
+//         return 'Outubro';
+//       case 11:
+//         return 'Novembro';
+//       case 12:
+//         return 'Dezembro';
+//       default:
+//         return '';
+//     }
+//   }
 
-  static String returnWeekDayByNumber(int number) {
-    switch (number) {
-      case 1:
-        return 'Segunda';
-      case 2:
-        return 'Terça';
-      case 3:
-        return 'Quarta';
-      case 4:
-        return 'Quinta';
-      case 5:
-        return 'Sexta';
-      case 6:
-        return 'Sábado';
-      case 7:
-        return 'Domingo';
-      default:
-        return '';
-    }
-  }
+//   static String returnWeekDayByNumber(int number) {
+//     switch (number) {
+//       case 1:
+//         return 'Segunda';
+//       case 2:
+//         return 'Terça';
+//       case 3:
+//         return 'Quarta';
+//       case 4:
+//         return 'Quinta';
+//       case 5:
+//         return 'Sexta';
+//       case 6:
+//         return 'Sábado';
+//       case 7:
+//         return 'Domingo';
+//       default:
+//         return '';
+//     }
+//   }
 
-  static String weekDayMonth(DateTime date) {
-    return '${returnWeekDayByNumber(
-      date.weekday,
-    )}, ${date.day} de ${getMonthByInt(
-      date.month,
-    )}';
-  }
-}
+//   static String weekDayMonth(DateTime date) {
+//     return '${returnWeekDayByNumber(
+//       date.weekday,
+//     )}, ${date.day} de ${getMonthByInt(
+//       date.month,
+//     )}';
+//   }
+// }