+++ lib/src/core/utils/validators/email_validator.dart
@@ -1,9 +1,0 @@
-class EmailValidator {
-  bool isValidEmail(String email) {
-    RegExp emailRegex = RegExp(
-      r'^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$',
-    );
-
-    return emailRegex.hasMatch(email);
-  }
-}