+++ lib/src/infra/drivers/i_connectivity_driver.dart
@@ -4,5 +4,5 @@ import '../../domain/enums/connectivity_status.dart';
 
 abstract class IConnectivityDriver {
   Future<bool> get isOnline;
-  StreamSubscription<ConnectivityStatus> get onConnectivityChanged;
+  StreamSubscription<List<ConnectivityStatus>> get onConnectivityChanged;
 }